보드프로젝트

+++ src/main/java/com/example/board/domain/dao/AttachDAO.java
@@ -11,16 +11,16 @@ import java.util.List;
 @RequiredArgsConstructor
 public class AttachDAO {
     private final AttachMapper attachMapper;
-
     public void register(AttachVO attachVO){
         attachMapper.insert(attachVO);
     }
-
     public void remove(String uuid){
+        if(uuid == null) {return;}
         attachMapper.delete(uuid);
     }
-
     public List<AttachVO> selectAllByBno(Long bno){
         return attachMapper.select(bno);
     }
+    public String getUUID(Long bno) { return attachMapper.getUUID(bno); }
+    public List<AttachVO> getOldFiles() { return attachMapper.getOldFiles(); }
 }
